fixed pacman interface and clarified board vs. game run interfaces
