+++ pacman/pacman_agents.py
@@ -6,7 +6,8 @@ class PacmanRandom:
     def __init__(self, vec, direction, board):
         self.board = board
         self.direction = direction
-        self.agent = vec
+        self.x = vec.x
+        self.y = vec.y
         self.moves = [
             vector(10, 0),
             vector(-10, 0),
@@ -14,19 +15,45 @@ class PacmanRandom:
             vector(0, -10)
         ]
 
-    def send_move_to_board(self):
-        return self.direction
-
     def move(self, valid):
         if valid:
-            self.agent.move(self.direction)
+            return True
         else:
             self.choose_direction()
 
     def choose_direction(self):
         self.direction = choice(self.moves)
 
+    def update(self, new_location):
+        self.x = new_location.x
+        self.y = new_location.y
 
+
+class PacmanBetterRandom:
+    def __init__(self, vec, direction, board):
+        self.board = board
+        self.direction = direction
+        self.x = vec.x
+        self.y = vec.y
+        self.moves = [
+            vector(10, 0),
+            vector(-10, 0),
+            vector(0, 10),
+            vector(0, -10)
+        ]
+
+    def move(self, valid):
+        if valid:
+            return True
+        else:
+            self.choose_move()
+
+
+    # def choose_move(self):
+    #
+    #
+    # def all_valid(self, position):
+    #     for i in self.moves:
