+++ pacman/pacman_driver.py
@@ -1,4 +1,4 @@
-from pacman.pacman_board import PacmanBoard
+from pacman.pacman_game import PacmanGame
 from pacman.pacman_agents import PacmanRandom
 from pacman.ghost_agents import GhostRandom
 from copy import deepcopy
@@ -14,6 +14,6 @@ ghostDir = vector(10, 0)
 ghosty = GhostRandom(ghost, ghostDir, deepcopy(tiles))
 ghostz = [ghosty]
 
-game = PacmanBoard(deepcopy(tiles), deepcopy(pacman), deepcopy(ghostz))
+game = PacmanGame(deepcopy(tiles), deepcopy(pacman), deepcopy(ghostz))
 game.game_setup()
